Refactor: 패키지 리네이밍(common -> command)에 따른 경로 업데이트
