+++ cash/src/main/java/com/mossy/boundedContext/cash/app/CashFacade.java
@@ -14,8 +14,8 @@ import com.mossy.boundedContext.cash.app.usecase.user.CashGetBalanceUseCase;
 import com.mossy.boundedContext.cash.app.usecase.user.CashGetLogsUseCase;
 import com.mossy.boundedContext.cash.app.usecase.user.CashGetWalletInfoUseCase;
 import com.mossy.boundedContext.cash.app.usecase.user.CashSyncUserUseCase;
-import com.mossy.boundedContext.cash.in.dto.common.CashSellerDto;
-import com.mossy.boundedContext.cash.in.dto.common.CashUserDto;
+import com.mossy.boundedContext.cash.in.dto.command.CashSellerDto;
+import com.mossy.boundedContext.cash.in.dto.command.CashUserDto;
 import com.mossy.boundedContext.cash.in.dto.request.CashHoldingRequestDto;
 import com.mossy.boundedContext.cash.in.dto.request.CashRefundRequestDto;
 import com.mossy.boundedContext.cash.in.dto.request.SellerBalanceRequestDto;
