+++ cash/src/main/java/com/mossy/boundedContext/cash/app/usecase/seller/CashCreateSellerWalletUseCase.java
@@ -3,7 +3,7 @@ package com.mossy.boundedContext.cash.app.usecase.seller;
 import com.mossy.boundedContext.cash.app.CashSupport;
 import com.mossy.boundedContext.cash.domain.seller.CashSeller;
 import com.mossy.boundedContext.cash.domain.seller.SellerWallet;
-import com.mossy.boundedContext.cash.in.dto.common.CashSellerDto;
+import com.mossy.boundedContext.cash.in.dto.command.CashSellerDto;
 import com.mossy.boundedContext.cash.out.seller.CashSellerRepository;
 import com.mossy.boundedContext.cash.out.seller.SellerWalletRepository;
 import lombok.RequiredArgsConstructor;
